refactor(logic): split fog_resolution into part-of concern files (#3290)
Phase 0 (#3290 C0) decomposition of the 545-line
packages/colonizethis_logic/lib/src/world/fog_resolution.dart into a thin
library root plus four part-of concern files:

- fog_resolution_explorer_spy_decay.dart (Explorer/Spy land fog decay +
  Spy reveal-timer maintenance)
- fog_resolution_province_ownership.dart (immediate visibility on province
  ownership transfer)
- fog_resolution_coastal_sea_zone.dart (coastal sea-zone full visibility)
- fog_resolution_distant_sea_zone.dart (distant sea-zone fog revert)

Pure structural extraction; no behaviour change. All public entry points
remain importable from fog_resolution.dart. Shared GP-visibility helpers
stay in the library root. Extends the land-province bucket-key guard
(tool/check_land_province_bucket_keys.dart) to cover the new part files so
its CI coverage is preserved.

Refs #3290
